+++ narrator.js
@@ -1,4 +1,5 @@
 "use strict";
+const MODULE = 'narrator-tools';
 class NarratorMenu extends FormApplication {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
@@ -11,27 +12,27 @@ class NarratorMenu extends FormApplication {
     }
     async getData(_options) {
         return {
-            FontSize: game.settings.get('narrator-tools', 'FontSize'),
-            WebFont: game.settings.get('narrator-tools', 'WebFont'),
-            TextColor: game.settings.get('narrator-tools', 'TextColor'),
-            TextShadow: game.settings.get('narrator-tools', 'TextShadow'),
-            TextCSS: game.settings.get('narrator-tools', 'TextCSS'),
-            Copy: game.settings.get('narrator-tools', 'Copy'),
-            Pause: game.settings.get('narrator-tools', 'Pause'),
-            DurationMultiplier: game.settings.get('narrator-tools', 'DurationMultiplier'),
-            BGColor: game.settings.get('narrator-tools', 'BGColor'),
-            BGImage: game.settings.get('narrator-tools', 'BGImage'),
-            NarrationStartPaused: game.settings.get('narrator-tools', 'NarrationStartPaused'),
-            MessageType: game.settings.get('narrator-tools', 'MessageType'),
+            FontSize: game.settings.get(MODULE, 'FontSize'),
+            WebFont: game.settings.get(MODULE, 'WebFont'),
+            TextColor: game.settings.get(MODULE, 'TextColor'),
+            TextShadow: game.settings.get(MODULE, 'TextShadow'),
+            TextCSS: game.settings.get(MODULE, 'TextCSS'),
+            Copy: game.settings.get(MODULE, 'Copy'),
+            Pause: game.settings.get(MODULE, 'Pause'),
+            DurationMultiplier: game.settings.get(MODULE, 'DurationMultiplier'),
+            BGColor: game.settings.get(MODULE, 'BGColor'),
+            BGImage: game.settings.get(MODULE, 'BGImage'),
+            NarrationStartPaused: game.settings.get(MODULE, 'NarrationStartPaused'),
+            MessageType: game.settings.get(MODULE, 'MessageType'),
             CHAT_MESSAGE_TYPES: {
                 0: 'Other',
                 1: 'Out of Character',
                 2: 'In Character',
             },
-            PERMScenery: game.settings.get('narrator-tools', 'PERMScenery'),
-            PERMDescribe: game.settings.get('narrator-tools', 'PERMDescribe'),
-            PERMNarrate: game.settings.get('narrator-tools', 'PERMNarrate'),
-            PERMAs: game.settings.get('narrator-tools', 'PERMAs'),
+            PERMScenery: game.settings.get(MODULE, 'PERMScenery'),
+            PERMDescribe: game.settings.get(MODULE, 'PERMDescribe'),
+            PERMNarrate: game.settings.get(MODULE, 'PERMNarrate'),
+            PERMAs: game.settings.get(MODULE, 'PERMAs'),
             USER_ROLES: {
                 0: game.i18n.localize('USER.RoleNone'),
                 1: game.i18n.localize('USER.RolePlayer'),
@@ -43,7 +44,7 @@ class NarratorMenu extends FormApplication {
     }
     async _updateObject(_event, formData) {
         for (let [k, v] of Object.entries(formData)) {
-            game.settings.set('narrator-tools', k, v);
+            game.settings.set(MODULE, k, v);
         }
         setTimeout(() => {
             NarratorTools._updateContentStyle();
@@ -57,14 +58,14 @@ const NarratorTools = {
     _chatMessage(message, content, chatData) {
         let commands = {};
         content = content.replace(/\n/g, '<br>');
-        if (game.user.role >= game.settings.get('narrator-tools', 'PERMAs')) {
+        if (game.user.role >= game.settings.get(MODULE, 'PERMAs')) {
             commands.as = new RegExp('^(?:\\/as$|\\/as ([^]*))', 'i');
         }
-        if (game.user.role >= game.settings.get('narrator-tools', 'PERMDescribe')) {
+        if (game.user.role >= game.settings.get(MODULE, 'PERMDescribe')) {
             commands.description = new RegExp('^\\/desc(?:ribe|ription|) ([^]*)', 'i');
             commands.notification = new RegExp('^\\/not(?:e|ify|ication) ([^]*)', 'i');
         }
-        if (game.user.role >= game.settings.get('narrator-tools', 'PERMNarrate')) {
+        if (game.user.role >= game.settings.get(MODULE, 'PERMNarrate')) {
             commands.narration = new RegExp('^\\/narrat(?:e|ion) ([^]*)', 'i');
         }
         let c, rgx, match;
@@ -93,7 +94,7 @@ const NarratorTools = {
     },
     _controller({ narration, scenery }) {
         this._updateScenery(scenery);
-        if (game.user.role >= game.settings.get('narrator-tools', 'PERMScenery')) {
+        if (game.user.role >= game.settings.get(MODULE, 'PERMScenery')) {
             const btn = $('.scene-control[data-tool=scenery]');
             if (btn) {
                 if (scenery)
@@ -117,7 +118,7 @@ const NarratorTools = {
                     let duration = this.messageDuration(this.sharedState.narration.message.length);
                     if (scroll > 20) {
                         const remaining = 1 - Number(this.elements.content[0].style.top.slice(0, -2)) / -scroll;
-                        const duration_multiplier = game.settings.get('narrator-tools', 'DurationMultiplier');
+                        const duration_multiplier = game.settings.get(MODULE, 'DurationMultiplier');
                         const scroll_duration = (duration - 500 - 4500 * duration_multiplier) * remaining;
                         const fun_scroll = () => {
                             this.elements.content.animate({ top: -scroll }, scroll_duration, 'linear');
@@ -145,7 +146,7 @@ const NarratorTools = {
                 clearTimeout(this._timeouts.narrationOpens);
                 this.elements.content[0].style.opacity = '0';
                 this.elements.content.stop();
-                this.elements.buttonCopy[0].style.display = game.settings.get('narrator-tools', 'Copy') ? '' : 'none';
+                this.elements.buttonCopy[0].style.display = game.settings.get(MODULE, 'Copy') ? '' : 'none';
                 this._timeouts.narrationOpens = +setTimeout(() => {
                     this.elements.content.html(narration.message);
                     this.elements.content[0].style.opacity = '1';
@@ -155,7 +156,7 @@ const NarratorTools = {
                     this.elements.buttons[0].style.opacity = '1';
                     this.elements.buttons[0].style.visibility = 'visible';
                     this.elements.buttons[0].style.top = `calc(50% + ${60 + height / 2}px)`;
-                    this._updateStopButton(game.settings.get('narrator-tools', 'NarrationStartPaused'));
+                    this._updateStopButton(game.settings.get(MODULE, 'NarrationStartPaused'));
                     this._timeouts.narrationOpens = 0;
                     Hooks.call('narration', narration);
                 }, 500);
@@ -180,7 +181,7 @@ const NarratorTools = {
         }
     },
     _createSceneryButton(control, html, data) {
-        const hasPerm = game.user.role >= game.settings.get('narrator-tools', 'PERMScenery');
+        const hasPerm = game.user.role >= game.settings.get(MODULE, 'PERMScenery');
         if (hasPerm) {
             const name = 'scenery';
             const title = game.i18n.localize('NT.ButtonTitle');
@@ -233,17 +234,17 @@ const NarratorTools = {
         }, 250);
     },
     _pause() {
-        const canScenery = game.user.role >= game.settings.get('narrator-tools', 'PERMScenery');
-        if (canScenery && game.settings.get('narrator-tools', 'Pause')) {
+        const canScenery = game.user.role >= game.settings.get(MODULE, 'PERMScenery');
+        if (canScenery && game.settings.get(MODULE, 'Pause')) {
             NarratorTools.scenery(game.paused);
         }
     },
     _preCreateChatMessage(chatMessage, options, user) {
-        if (game.user.role >= game.settings.get('narrator-tools', 'PERMAs') && this.character) {
+        if (game.user.role >= game.settings.get(MODULE, 'PERMAs') && this.character) {
             let chatData = {};
-            chatData.type = game.settings.get('narrator-tools', 'MessageType');
+            chatData.type = game.settings.get(MODULE, 'MessageType');
             chatData.speaker = { alias: this.character };
-            chatMessage.data.update(chatData);
+            chatMessage.update(chatData);
         }
     },
     _ready() {
@@ -264,7 +265,7 @@ const NarratorTools = {
         this._updateBGImage();
         this._fitSidebar();
         $('body').append(this._element);
-        this.isNarrator = game.user.hasPermission('SETTINGS_MODIFY') && game.user.role >= game.settings.get('narrator-tools', 'PERMNarrate');
+        this.isNarrator = game.user.hasPermission('SETTINGS_MODIFY') && game.user.role >= game.settings.get(MODULE, 'PERMNarrate');
         this._menu = new ContextMenuNT({
             theme: 'default',
             items: [
@@ -308,12 +309,12 @@ const NarratorTools = {
             this.elements.buttonPause[0].style.display = 'none';
             this.elements.buttonClose[0].style.display = 'none';
         }
-        this._loadFont(game.settings.get('narrator-tools', 'WebFont'));
+        this._loadFont(game.settings.get(MODULE, 'WebFont'));
         this._updateContentStyle();
-        this._controller(game.settings.get('narrator-tools', 'sharedState'));
+        this._controller(game.settings.get(MODULE, 'sharedState'));
         this._pause();
         document.addEventListener('contextmenu', (ev) => {
-            if (ev.target.classList.contains('editor-content') || $(ev.target).parents('div.editor-content').length) {
+            if (ev.target.classList.contains('journal-entry-pages') || $(ev.target).parents('div.journal-entry-pages').length) {
                 const time = this._menu.isOpen() ? 100 : 0;
                 this._menu.hide();
                 setTimeout(() => {
@@ -324,7 +325,11 @@ const NarratorTools = {
         document.addEventListener('click', () => NarratorTools._menu.hide());
     },
     _setup() {
-        game.settings.register('narrator-tools', 'sharedState', {
+        this._registerKeybindings();
+        this._registerGameSettings();
+    },
+    _registerGameSettings() {
+        game.settings.register(MODULE, 'sharedState', {
             name: 'Shared State',
             scope: 'world',
             config: false,
@@ -340,21 +345,21 @@ const NarratorTools = {
             },
             onChange: (newState) => this._controller(newState),
         });
-        game.settings.registerMenu('narrator-tools', 'settingsMenu', {
+        game.settings.registerMenu(MODULE, 'settingsMenu', {
             name: game.i18n.localize('SETTINGS.Configure'),
             label: game.i18n.localize('SCENES.Configure'),
             icon: 'fas fa-adjust',
             type: NarratorMenu,
             restricted: true,
         });
-        game.settings.register('narrator-tools', 'FontSize', {
+        game.settings.register(MODULE, 'FontSize', {
             name: 'Font Size',
             scope: 'world',
             config: false,
             default: '',
             type: String,
         });
-        game.settings.register('narrator-tools', 'WebFont', {
+        game.settings.register(MODULE, 'WebFont', {
             name: 'Web Font',
             scope: 'world',
             config: false,
@@ -362,49 +367,49 @@ const NarratorTools = {
             type: String,
             onChange: (value) => NarratorTools._loadFont(value),
         });
-        game.settings.register('narrator-tools', 'TextColor', {
+        game.settings.register(MODULE, 'TextColor', {
             name: 'Text Color',
             scope: 'world',
             config: false,
             default: '',
             type: String,
         });
-        game.settings.register('narrator-tools', 'TextShadow', {
+        game.settings.register(MODULE, 'TextShadow', {
             name: 'Text Shadow',
             scope: 'world',
             config: false,
             default: '',
             type: String,
         });
-        game.settings.register('narrator-tools', 'TextCSS', {
+        game.settings.register(MODULE, 'TextCSS', {
             name: 'TextCSS',
             scope: 'world',
             config: false,
             default: '',
             type: String,
         });
-        game.settings.register('narrator-tools', 'Copy', {
+        game.settings.register(MODULE, 'Copy', {
             name: 'Copy',
             scope: 'world',
             config: false,
             default: false,
             type: Boolean,
         });
-        game.settings.register('narrator-tools', 'Pause', {
+        game.settings.register(MODULE, 'Pause', {
             name: 'Pause',
             scope: 'world',
             config: false,
             default: false,
             type: Boolean,
         });
-        game.settings.register('narrator-tools', 'DurationMultiplier', {
+        game.settings.register(MODULE, 'DurationMultiplier', {
             name: 'Duration Multiplier',
             scope: 'world',
             config: false,
             default: 1,
             type: Number,
         });
-        game.settings.register('narrator-tools', 'BGColor', {
+        game.settings.register(MODULE, 'BGColor', {
             name: 'Background Color',
             scope: 'world',
             config: false,
@@ -412,7 +417,7 @@ const NarratorTools = {
             type: String,
             onChange: (color) => NarratorTools._updateBGColor(color),
         });
-        game.settings.register('narrator-tools', 'BGImage', {
+        game.settings.register(MODULE, 'BGImage', {
             name: 'Background Color',
             scope: 'world',
             config: false,
@@ -420,42 +425,42 @@ const NarratorTools = {
             type: String,
             onChange: (filePath) => NarratorTools._updateBGImage(filePath),
         });
-        game.settings.register('narrator-tools', 'NarrationStartPaused', {
+        game.settings.register(MODULE, 'NarrationStartPaused', {
             name: 'Start the Narration Paused',
             scope: 'world',
             config: false,
             default: false,
             type: Boolean,
         });
-        game.settings.register('narrator-tools', 'MessageType', {
+        game.settings.register(MODULE, 'MessageType', {
             name: 'Narration Message Type',
             scope: 'world',
             config: false,
             default: CONST.CHAT_MESSAGE_TYPES.OTHER,
             type: Number,
         });
-        game.settings.register('narrator-tools', 'PERMScenery', {
+        game.settings.register(MODULE, 'PERMScenery', {
             name: 'Permission Required to set the Scenery',
             scope: 'world',
             config: false,
             default: CONST.USER_ROLES.GAMEMASTER,
             type: Number,
         });
-        game.settings.register('narrator-tools', 'PERMDescribe', {
+        game.settings.register(MODULE, 'PERMDescribe', {
             name: 'Permission Required to /describe and /note',
             scope: 'world',
             config: false,
             default: CONST.USER_ROLES.GAMEMASTER,
             type: Number,
         });
-        game.settings.register('narrator-tools', 'PERMNarrate', {
+        game.settings.register(MODULE, 'PERMNarrate', {
             name: 'Permission Required to /narrate',
             scope: 'world',
             config: false,
             default: CONST.USER_ROLES.GAMEMASTER,
             type: Number,
         });
-        game.settings.register('narrator-tools', 'PERMAs', {
+        game.settings.register(MODULE, 'PERMAs', {
             name: 'Permission Required to /as',
             scope: 'world',
             config: false,
@@ -482,7 +487,7 @@ const NarratorTools = {
         commands[data.command]();
     },
     _renderChatMessage(message, html, data) {
-        const type = message.getFlag('narrator-tools', 'type');
+        const type = message.getFlag(MODULE, 'type');
         if (type) {
             html.find('.message-sender').text('');
             html.find('.message-metadata')[0].style.display = 'none';
@@ -517,29 +522,29 @@ const NarratorTools = {
         }
     },
     _updateBGColor(color) {
-        color = color ?? (game.settings.get('narrator-tools', 'BGColor') || '#000000');
+        color = color ?? (game.settings.get(MODULE, 'BGColor') || '#000000');
         this.elements.frameBG[0].style.boxShadow = `inset 0 0 2000px 100px ${color}`;
         this.elements.BG[0].style.background = `linear-gradient(transparent 0%, ${color}a8 40%, ${color}a8 60%, transparent 100%)`;
     },
     _updateBGImage(filePath) {
-        filePath = filePath ?? game.settings.get('narrator-tools', 'BGImage') ?? '';
+        filePath = filePath ?? game.settings.get(MODULE, 'BGImage') ?? '';
         if (filePath) {
             this.elements.frameBG[0].style.background = `url(${filePath})`;
             this.elements.frameBG[0].style.backgroundSize = '100% 100%';
         }
     },
     _updateContentStyle() {
-        const style = game.settings.get('narrator-tools', 'TextCSS');
+        const style = game.settings.get(MODULE, 'TextCSS');
         if (style) {
             const opacity = this.elements.content[0].style.opacity;
             this.elements.content[0].style = style;
             this.elements.content[0].style.opacity = opacity;
             return;
         }
-        this.elements.content[0].style.fontFamily = `${game.settings.get('narrator-tools', 'WebFont')}` ? 'NTCustomFont' : '';
-        this.elements.content[0].style.fontSize = `${game.settings.get('narrator-tools', 'FontSize')}`;
-        this.elements.content[0].style.color = `${game.settings.get('narrator-tools', 'TextColor')}`;
-        this.elements.content[0].style.textShadow = `${game.settings.get('narrator-tools', 'TextShadow')}`;
+        this.elements.content[0].style.fontFamily = `${game.settings.get(MODULE, 'WebFont')}` ? 'NTCustomFont' : '';
+        this.elements.content[0].style.fontSize = `${game.settings.get(MODULE, 'FontSize')}`;
+        this.elements.content[0].style.color = `${game.settings.get(MODULE, 'TextColor')}`;
+        this.elements.content[0].style.textShadow = `${game.settings.get(MODULE, 'TextShadow')}`;
     },
     _updateScenery(scenery) {
         if (!scenery)
@@ -549,6 +554,21 @@ const NarratorTools = {
             return;
         this.elements.frameBG[0].style.opacity = new_state;
         this.elements.sidebarBG[0].style.opacity = new_state;
+    },
+    _registerKeybindings() {
+        game.keybindings.register(MODULE, 'toogleScenery', {
+            name: 'Toggle Scenery Control',
+            hint: 'Turns the scenery on/off on command.',
+            editable: [
+                {
+                    key: 'F1',
+                    modifiers: [],
+                },
+            ],
+            onDown: () => NarratorTools.scenery(),
+            restricted: true,
+            precedence: CONST.KEYBINDING_PRECEDENCE.NORMAL,
+        });
     },
     messagesQueue: [],
     chatMessage: {
@@ -568,9 +588,9 @@ const NarratorTools = {
         },
     },
     createChatMessage(type, message, options = {}) {
-        if (type == 'narration' && !game.user.role >= game.settings.get('narrator-tools', 'PERMNarrate'))
+        if (type == 'narration' && !game.user.role >= game.settings.get(MODULE, 'PERMNarrate'))
             return;
-        else if (!game.user.role >= game.settings.get('narrator-tools', 'PERMDescribe'))
+        else if (!game.user.role >= game.settings.get(MODULE, 'PERMDescribe'))
             return;
         message = message.replace(/\\n/g, '<br>');
         let chatData = {
@@ -580,7 +600,7 @@ const NarratorTools = {
                     type: type,
                 },
             },
-            type: game.settings.get('narrator-tools', 'MessageType'),
+            type: game.settings.get(MODULE, 'MessageType'),
             speaker: {
                 alias: game.i18n.localize('NT.Narrator'),
                 scene: game.user.viewedScene,
@@ -610,7 +630,7 @@ const NarratorTools = {
                 id: this.sharedState.narration.id + 1,
                 display: true,
                 message: messageStripped,
-                paused: game.settings.get('narrator-tools', 'NarrationStartPaused'),
+                paused: game.settings.get(MODULE, 'NarrationStartPaused'),
             };
             this.sharedState.narration = state;
             ChatMessage.create(chatData, {});
@@ -621,10 +641,10 @@ const NarratorTools = {
     elements: {},
     isNarrator: false,
     messageDuration(length) {
-        return (Math.clamped(2000, length * 80, 20000) + 3000) * game.settings.get('narrator-tools', 'DurationMultiplier') + 500;
+        return (Math.clamped(2000, length * 80, 20000) + 3000) * game.settings.get(MODULE, 'DurationMultiplier') + 500;
     },
     scenery(state) {
-        if (game.user.role >= game.settings.get('narrator-tools', 'PERMScenery')) {
+        if (game.user.role >= game.settings.get(MODULE, 'PERMScenery')) {
             if (!game.user.hasPermission('SETTINGS_MODIFY'))
                 ui.notifications.error(game.i18n.localize('NT.CantModifySettings'));
             else
@@ -633,18 +653,18 @@ const NarratorTools = {
     },
     sharedState: {
         get narration() {
-            return game.settings.get('narrator-tools', 'sharedState').narration;
+            return game.settings.get(MODULE, 'sharedState').narration;
         },
         set narration(state) {
-            const sharedState = { ...game.settings.get('narrator-tools', 'sharedState'), narration: state };
-            game.settings.set('narrator-tools', 'sharedState', sharedState);
+            const sharedState = { ...game.settings.get(MODULE, 'sharedState'), narration: state };
+            game.settings.set(MODULE, 'sharedState', sharedState);
         },
         get scenery() {
-            return game.settings.get('narrator-tools', 'sharedState').scenery;
+            return game.settings.get(MODULE, 'sharedState').scenery;
         },
         set scenery(state) {
-            const sharedState = { ...game.settings.get('narrator-tools', 'sharedState'), scenery: state };
-            game.settings.set('narrator-tools', 'sharedState', sharedState);
+            const sharedState = { ...game.settings.get(MODULE, 'sharedState'), scenery: state };
+            game.settings.set(MODULE, 'sharedState', sharedState);
         },
     },
 };